Moving things around and refactoring
